+++ CHANGELOG.md
@@ -42,9 +42,12 @@ Section Order:
 
 <!-- Your changes go here -->
 
+## [3.2.1] - 2026-01-05
+
 ### Changed
 
 - Update pending SRP request number in the menu badge when actioning requests
+- Translations updated
 
 ## [3.2.0] - 2025-12-02
 
@@ -1594,11 +1597,12 @@ python manage.py aasrp_update_db_relations
 [3.1.1]: https://github.com/ppfeufer/aa-srp/compare/v3.1.0...v3.1.1 "v3.1.1"
 [3.1.2]: https://github.com/ppfeufer/aa-srp/compare/v3.1.1...v3.1.2 "v3.1.2"
 [3.2.0]: https://github.com/ppfeufer/aa-srp/compare/v3.1.2...v3.2.0 "v3.2.0"
+[3.2.1]: https://github.com/ppfeufer/aa-srp/compare/v3.2.0...v3.2.1 "v3.2.1"
 [aa discord notify]: https://gitlab.com/ErikKalkoken/aa-discordnotify "AA Discord Notify"
 [aa fleet pings]: https://github.com/ppfeufer/aa-fleetpings "AA Fleet Pings"
 [aa-discordbot]: https://github.com/pvyParts/allianceauth-discordbot "AA-Discordbot"
 [evetools killboard]: https://kb.evetools.org/ "EveTools Killboard"
-[in development]: https://github.com/ppfeufer/aa-srp/compare/v3.2.0...HEAD "In Development"
+[in development]: https://github.com/ppfeufer/aa-srp/compare/v3.2.1...HEAD "In Development"
 [keep a changelog]: http://keepachangelog.com/ "Keep a Changelog"
 [semantic versioning]: http://semver.org/ "Semantic Versioning"
 [tooltip: change srp payout amount]: https://raw.githubusercontent.com/ppfeufer/aa-srp/master/docs/images/tooltip-change-srp-payout-amount.png "Tooltip: Change SRP Payout Amount"